requirements.api.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.2
numpy>=1.21.0
pandas>=1.3.0
./edapi/edapi
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
